ecommerce/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
Encoder==1.1
idna==3.7
kombu==5.4.0
mailtrap==2.0.1
mysqlclient==2.2.4
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
